Move PostNavigation to component folder
